change when ECA fires; remove code that prevents page jumping and put in readMe as suggestion; tidy up setAnimationProperties function:
Previously, ECA fired on page load event. Now it fires on DCL.
Having it fire on load was mainly to prevent page jumping and
to prevent situations where content (images, fonts) might not
be present yet. However, it wasn't performant, especially on
slow networks, so I changed it.

I also removed code that fixed page jumping and made it a
mere suggestion in the readME.

Finally, I cleaned up the setAnimationProperties function a bit
so it would be easier to read.