refactor, add new functionality to, and better comment code; change animate classes to eca-animate...; move props off HTML elements and place into own namespace; add new elem/elemGroup prop atRemoveAnimationPoint to standardize how eca-animate class is removed from elem(s); expand functionality of remove-animation-when-not-in-view option; remove capture option and change to listenOnce; add throttle limit options for scroll and resize events;
I removed the following functions: getLetters, getElementsToArray,
setAnimationProperties, ifElementsNeedAnimating. getLetters and
getElementsToArray were no longer needed (ifElements... was
unnecessary, and the elements were already an array with the new
constructor I wrote). I moved code for setAnimationProperties
into its own constructor, since it concerned
each element group we added inside the readyElementsForAnimation
function.

Regarding the readyElementsForAnimation function, I massively simplified
the code. All functions private to that function were moved onto the
eca.animatable prop, and the logic for the function was overall
simplified--e.g., all the gathering and sorting for text elems and regular
elems is done in one for loop now.

I refactored the following functions: elementInView, getMaxTriggerOffset,
wrapText, setStyles, setListeners, getEventTypes, requestAnimationUpdate,
updateAnimations, animate--basically, almost every function.

The main work went into refactoring and fixing the wrapText function.
The problem the function had was wrapping text with HTML tags inside.
While it still does not handle every use case, it handles more now. Furthermore,
it's easier to follow the logic (the logic is iterative instead of recursive now).

The rest of the functions were simplified slightly (logic simplified, better
var names, easier to read conditionals), or changed to reflect new
logic the app allows for, as with the animate function
(e.g. new conditional logic for the
remove-animation-when-not-in-view function and animate-all-on-first-sight
options).

Regarding options, I removed the capture option, as it didn't do anything
(I only fire the handler when e.target is e.currentTarget, so capturing and
bubbling are no longer relevant). The capture option is now the listenOnce
option, which can be used if the remove-animation-when-not-in-view option is
not set to remove the listener after it fires once.

As for the remove-animation-when-not-in-view option, when set to 'true,' it now supports
removing the eca-animated class when the elem is both above or below
the viewport. The user can also now set it to either 'above' or 'below' to only
remove the element when it's only above or below the viewport, but not
the other direction.

Finally, I changed the name of a few functions, such as setAnimationDelay
which is now setAnimationDelayProp, to better reflect what they do
(e.g., in this case, saving a property on our ecaElemProps obj and not
actually writing to the style prop on the elem).